pyproject.toml
[tool.poetry]
name = "mason-framework"
version = "0.0.0"
description = "Node based building blocks."
authors = []
packages = [
  { include = "mason"}
]

[tool.poetry.scripts]
mason = 'mason.cli:mason_cli'

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^19.3.0"
protobuf = "^3.11.3"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
aioconsole = "^0.1.16"
pylint = "^2.4.4"
tox = "^3.14.6"
coverage = "^5.0.4"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
pytest-asyncio = "^0.11.0"
mock = "^4.0.2"
asyncmock = "^0.4.2"
pytest-sugar = "^0.9.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
